Tried to implement url feature; leaving for future update
A url button next to a tab group would display the tabs that are in that group. It would display the each tab's icon, name, and url.
The urlButton's dropdown is working correctly, but I was unable to display the icon. Also, the information is not displayed correctly in the dropdown.