Fixed tab color problem
Process to produce problem before this commit:
1. Open a tab
2. Color it a color (red)
3. Save it (name: a)
4. Color the same tab (green)
5. Save it (name: b)
6. With green tab still being active, click on tab buttons in popup (a and b) to launch tabs
The result would be that the "a" tab would now be colored green instead of staying red.

Problem:
Once the tab is updated, it would look at the current tab's color and then assign it to the tab that is being updated, which would be the "a" tab.

Solution:
Check if the tab id of the tab being updated is the same as the current tab's id. If it is not, do not update (means that another tab is being updated/refreshed). If it is, then that means the current tab is being refreshed so then it would correctly set the tab's color from saveColor for it to persist through tab refresh.